Renamed KadotEngine -> Kodot, added Input to io pkg
